added test for wait
