--- internal_testsuites/golang/testsuite/startosis_request_wait_assert_test/startsois_validation_failure_invalid_port_id_request_test.go
@@ -1,44 +1,0 @@
-package startosis_request_wait_assert_test
-
-import (
-	"context"
-	"github.com/kurtosis-tech/kurtosis-cli/golang_internal_testsuite/test_helpers"
-	"github.com/stretchr/testify/require"
-	"testing"
-)
-
-const (
-	requestInvalidPortIDTest = "starlark-request-invalid-portid"
-	assertFailScript         = `
-def run(plan):
-	service_config = ServiceConfig(
-		image = "mendhak/http-https-echo:26",
-		ports = {
-			"http-port": PortSpec(number = 8080, transport_protocol = "TCP")
-		}
-	)
-
-	plan.add_service(name = "web-server", config = service_config)
-	get_recipe = GetHttpRequestRecipe(
-		port_id = "http-port",
-		endpoint = "?input=foo/bar",
-		extract = {
-			"exploded-slash": ".query.input | split(\"/\") | .[1]"
-		}
-	)
-	response = plan.wait(recipe=get_recipe, field="code", assertion="==", target_value=200, interval="100ms", timeout="30s", service_name="web-server")
-	plan.assert(response["code"], "!=", 200)
-
-	# dumb test to validate we can pass 2 runtime values here
-	plan.assert(response["code"], "==", response["code"])
-`
-)
-
-func TestStartosis_AssertFail(t *testing.T) {
-	ctx := context.Background()
-	runResult := test_helpers.SetupSimpleEnclaveAndRunScript(t, ctx, assertFailTestName, assertFailScript)
-
-	require.Nil(t, runResult.InterpretationError, "Unexpected interpretation error")
-	require.Empty(t, runResult.ValidationErrors, "Unexpected validation error")
-	require.NotEmpty(t, runResult.ExecutionError, "Expected execution error coming from assert fail")
-}
